fix(epsilon): check for float infinity
float should be used instead of double, as the value is stored as a float

+++ epsilon/apps/statistics/categorical/data/store_controller.cpp
@@ -67,7 +67,7 @@ bool StoreController::textFieldShouldFinishEditing(
 
 bool StoreController::textFieldDidFinishEditing(
     Escher::AbstractTextField* textField, Ion::Events::Event event) {
-  double p = ParseInputFloatValue<double>(textField->draftText());
+  float p = ParseInputFloatValue<float>(textField->draftText());
   if (HasUndefinedValue(p)) {
     return false;
   }
